pyproject.toml
[tool.poetry]
name = "mlops-proj-cat-dog"
version = "0.1.0"
description = ""
authors = ["BedPled <vishniakov.ae@students.dvfu.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.26.3"
scikit-learn = "^1.3.2"
scikit-image = "^0.22.0"
mlflow = "^2.9.2"
pandas = "^2.1.4"
hydra-core = "^1.3.2"
dvc = "^3.38.1"
tensorflow-io-gcs-filesystem = "~0.31.0"
tensorflow = "~2.15.0"
keras = "~2.15.0"
tensorflow-estimator = "~2.15.0"
tensorflow-intel = "~2.15.0"
dvc-gdrive = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 130
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
'''

[tool.flake8]
max-line-length = 130
extend-ignore = ["D203", "E203", "E251", "E266", "E302", "E305", "E401", "E402", "E501", "F401", "F403", "W503"]
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10

[tool.isort]
atomic = true
profile = "black"
line_length = 130
skip_gitignore = true
